requirements.txt
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg2-binary==2.9.11
Pillow==12.2.0
qrcode[pil]==8.2
django-environ==0.13.0
gunicorn==25.3.0
whitenoise==6.12.0
